refactor: type-safe MemoryMap regions and misc core cleanup
- MemoryMap API now accepts `impl Into<RegionId>` so callers can pass
  typed enum variants instead of bare u8 constants
- Machine region IDs migrated from `mod region` constants to #[repr(u8)]
  enums with associated u8 constants for match-arm dispatch
- Simplify M6809/M6800 halt state: replace boxed ExecState::Halted with
  a plain `halted: bool` flag
- Fix bus_split! macro to not expand caller metavariables inside unsafe
- Remove unused variables in m6809_long_branch_test
- Various device/CPU import and signature cleanups

+++ core/src/core/debug.rs
@@ -76,7 +76,7 @@ pub trait BusDebug {
     /// Write to a device register by device index and register offset.
     /// Device indices match the order returned by `devices()`.
     /// Only `#[debug_device]` fields respond; CPU indices are ignored.
-    fn write_device_register(&mut self, _device_index: usize, _offset: u8, _data: u8) {}
+    fn write_device_register(&mut self, _device_index: usize, _offset: u16, _data: u8) {}
 
     /// Reset an individual device by index.
     /// Device indices match the order returned by `devices()`.
